Rename project and clean up codebase

+++ src/utils/bot.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import contextlib
 import re
 from typing import TYPE_CHECKING
 
@@ -17,6 +18,10 @@ async def download_file(message: Message) -> str:
     :param message: Message object.
     :return: Path to the downloaded file.
     """
+    if not message.bot:
+        msg = "There is no bot instance"
+        raise Exception(msg)
+
     if message.document:
         file_id = message.document.file_id
     elif message.photo:
@@ -26,13 +31,16 @@ async def download_file(message: Message) -> str:
     elif message.audio:
         file_id = message.audio.file_id
     else:
-        raise Exception("Unsupported file type.")
+        msg = "Unsupported file type."
+        raise Exception(msg)
 
     temp_file_path = f"temp_{file_id}"
     file = await message.bot.get_file(file_id)
-    if not file:
-        raise Exception("Only files up to 50MB are supported.")
     file_path = file.file_path
+    if not file_path:
+        msg = "Only files up to 50MB are supported."
+        raise Exception(msg)
+
     await message.bot.download_file(file_path=file_path, destination=temp_file_path)
     return temp_file_path
 
@@ -43,10 +51,12 @@ async def delete_file(file_path: str) -> None:
 
 async def delete_fsm_message(message_id: int, message: Message) -> None:
     """Deletes message by message_id that stores in FSM-data."""
-    try:
+    if not message.bot:
+        msg = "There is no bot instance"
+        raise Exception(msg)
+
+    with contextlib.suppress(TelegramBadRequest):
         await message.bot.delete_message(chat_id=message.chat.id, message_id=message_id)
-    except TelegramBadRequest:
-        pass
 
 
 def escape_markdown_v2(text: str) -> str:
